Fix model access errors unhelpful
In order to make errors more descriptive when things go wrong, this
commit adds a few error types to handle api interactions more gracefully
when things go wrong. Specifically, it currently handles when a model
can't be found or is innaccesible with the api token provided. Also when
an api token isn't provided, instead of an unhelpful NoMethodError, we
let end users know the token wasn't provided or isn't valid.