Prototyping third party gems (#451)
Demonstration of using third party cogs and binding them into a workflow.

This adds a new syntax `use`, in which you declare the cog you want to use and the gem it is from. `use` will ensure the gem is required and then bind the cog classes into the registry.

For testing, I've set up a very simple demo gem, and a standalone "project" that contains a roast workflow with its own gemfile. This ensures that gem loading works properly in projects using Roast.

A workflow can _also_ define workflow specific cogs. If `use` does not have the `from:` parameter passed, the workflow will look in the cogs directory for that given workflow (conventionally, the `cogs/` dir in the same directory as the workflow file).