Add DSL tests and minitest_dsl rake task (#459)
# Add DSL Testing Infrastructure and Example Tests

There was a small bug in cog_registry that got missed because we don't run functional tests (or any tests really) on the new DSL prototypes.

## Changes

This PR adds a bare-bones testing infrastructure for the DSL workflows and integrates it into the CI pipeline.

I copied the basic of the functional test infrastructure that @dersam set up for legacy roast functional tests. I opted not to try to re-use that existing class because it bakes in some legacy-specific workflow concepts (which is fine).

I did not re-implement the VCR / Cassette functionality yet. We'll want that for sure once we start running functional tests on workflows that use LLMs, but at the moment we don't need it to get started running this test suite in CI.

### Key changes include

1. Added a new `minitest_dsl` Rake task to run DSL-specific tests
2. Updated the CI workflow to execute the new DSL tests
3. Created a DSL-specific test support infrastructure with sandbox capabilities
4. Added functional tests for example DSL workflows (prototype.rb, scoped_executors.rb, step_communication.rb)
5. Removed outdated DSL examples (simple.rb, less_simple.rb)