Fix circular dependency coming from cli/ui (#461)
# Fix circular dependency coming from cli/ui

There's a circular dependency issue in the [cli/ui gem](https://github.com/Shopify/cli-ui/issues/607) that is producing a bunch of annoying warnings whenever we run tests in this project.

This PR bumps the cli-ui version to pull temporarily from the cli-ui commit with the fix. When that merges we will bump the version again.