Add skip! and fail! control flow methods to cogs, and bang- and question- versions of output accessors (#473)
This PR include two related concepts that are hard to separate in terms of implementation. I think we should discuss them together.

# Introduce `skip!`, and `fail!` methods for cog flow control

## skip!

Sometimes, you may have a reason to conditionally skip a cog, based on the output of a previous cog. The `skip!` method available in the cog's input proc will do this.

For instance, if you have a step that fetches some data that could be in several formats, you may need to transform it in some way, but only if it is in one of its possible formats. You could define a workflow like this in a simple and clean way:

(I'm using a pseudo-code cog named `cog` as a stand-in for any kind of cog; it'd doesn't matter to the example)

```
cog(:fetch_data) { "... fetches some data..." }

cog(:clean_data) do
  skip! if cog(:fetch_data).clean? 
  "... cleans the data..."
end
```

## fail!

You may also have a case where, because of some aspect of a prior cog's output (not necessarily a failure on its part, though), you don't have a valid state in which to continue the workflow. You could, of course, just raise any old exception, but having a `fail!` method that allows you to communicate to the workflow manager than the failure was sort of intentional allows us to better handle this kind of failure (we definitely don't need to show a stack trace, for example).

For instance, if you have a step that fetches data from an API, you might want to gracefully fail the workflow if the API returns a 404. This is a case you expect to encounter sometimes, though, so you want Roast to handle this kind of failure more gracefully. (TBD, though, what that 'gracefully' will actually look like, in terms of process exit code, what we print to the console, etc.) A workflow that uses this method might look something like this example.

```
cog(:fetch_api_data) { "... fetches some data from a web api..." }

cog(:process_data) do
  fail! unless cog(:fetch_api_data).response.status == 200
  " ... does something with the data ..."
end
```

# Add bang, non-bang, and question versions of cog output accessors

In a world where cogs can be skipped, and/or where cogs can fail but the workflow proceed, it becomes possible, in a subsequent cog's input proc, for the user to request the output of a previous cog that did not complete successfully, or to want to know if a previous cog did or did not run successfully

The current implementation of the cog output accessor methods (what I'm calling the cog name method you call inside a cog's input proc to get a prior cog's output) will raise an exception if the cog did not run and successfully produce output.

I think it's ugly and verbose to expect the user to rescue these exceptions in the input proc, especially if they need to do it a bunch. So, I've added bang and non-bang versions of the output accessor methods, along with a question-mark version that just returns a boolean if the cog's output is / is not present.

Pros:
* clean, clear syntax that is good Ruby style
* simple for user to decide "i want this to throw if the cog didn't run successfully" or "I want to handle the case where the cog didn't run successfully myself" or "I want to ask if this cog ran, but I don't need the output right now"
* `cog?(:name)` is clearer and more idiomatic that calling `cog(:name).nil?` / `cog(:name).present?` as a way to ask "did this cog run successfully?"

Cons:
* most of the time, the user probably wants the bang version of the method (though tbh I'm less sure now that this is true than when I first raised this PR)
* maybe the user still wants some of the output access exceptions to be raised (like, give me nil if the cog was skipped, but raise if the cog failed)

When I was working on this, I thought I was going to kind of hate having to use the bang methods most of the time, but when I actually updated the prototype workflows, I found out that I actually liked it. Curious what you all think.

(Also, for the second con, if the user wants some of the exceptions to still be raised, they can just use the bang method and rescue the ones that want to. I don't think we need a custom mechanism for that in the DSL).


UPDATE: The `CogDoesNotExistError` will *always* raise even with the non-bang or question-mark versions of the output accessor method. In such a case, the workflow is actually invalid, in a way that we would ideally be able to catch at compile time (but can't because the problematic statement is inside a cog input block that we can't evaluate until runtime).