Add doc comments to `cmd` cog config (#475)
This PR adds reasonable comprehensive user-facing documentation comments to the `cmd` cog's Config class`. These comments will show up in the user's IDE when they are developing a workflow, making discovery of configuration options for this cog, and information on what these options do and how to use the readily available.

Eventually, we should ship similar doc comments for all cogs, and for the execute and cog_input contexts as well, once the framework settles down. This PR is a first demonstration of what doing this could look like.

---

## Examples

Mousing over the `cmd` method in the `config` block, in RubyMine:
![Screenshot 2025-10-21 at 1.42.23 PM.png](https://app.graphite.dev/user-attachments/assets/65906ab2-ade5-42eb-aa31-5e44ca35e417.png)

Mousing over the `cmd` method in the `config` block, in RubyMine (continued):
![Screenshot 2025-10-21 at 1.42.29 PM.png](https://app.graphite.dev/user-attachments/assets/a41f60fb-265c-4bf9-94ed-71b9dd37bbb1.png)

Mousing over the `print_all!` method in the `cmd` cog's config block, in RubyMine:
![Screenshot 2025-10-21 at 1.42.35 PM.png](https://app.graphite.dev/user-attachments/assets/908f94ea-b880-45f3-99e3-e8a9ec9e347a.png)