Add `reduce` output accessor method to reduce results of running a `map` cog (#483)
This PR adds `reduce` output accessor method that mirrors the functionality of `collect` and `from` in downstack PRs.

The `reduce` output accessor has generally the same semantics as the normal Ruby `reduce` method on an iterable. You give it an optional initial_value, and a block that takes an accumulator as argument. The return value of the block becomes the value of the accumulator passed to the block the next time it is called, and the return value of the block the final time it is invoked becomes the return value of `reduce` itself.

There is one key semantic difference. Ruby's normal reduce, called on an iterable of objects, takes a block that accepts two arguments, the accumulator and each object from the iterable in turn. The `reduce` output accessor here only passes one argument, the accumulator, to its block. Instead, the block is executed __in the context of each execution scope from the map being reduced__. Thus, no second 'item' argument is necessary.