Add an `outputs` declaration to an `execute` block to set a custom 'return value' (#488)
Adds an optional `outputs` declaration to `execute` that allows you to specify a custom 'return value' for that scope when called as a subroutine. 

This lets you override the default 'return value' of the output of the final cog

Example:

```
execute(:sub) do
  cog(:one) { ... }
  cog(:two) { ... }
  cog(:three) { ... }
  outputs { cog(:two) }
end

execute do
  cog(:main) do
    val = from(call(:sub)) # now gives you the output of cog(:two) instead of cog(:three)
  end
end
```

The block given to `outputs` can return any value; it does not have to be the output of a cog.