Add optional index parameter when execution scopes are invoked with `map`
This PR adds an optional third parameter `index` to a cog_input_proc when it is
invoked inside an `execute` block that has been invoked by map. This provides functionality
equivalent to Ruby's `.map.with_index` method, and allows cogs running inside that scope to know
their index in the source enumerable.

The value of `index` will always be 0 when the scope is not invoked by `map`
(e.g., if it instead invoked by `call`). I considered having the value in this case be `nil`, but
I didn't want users to have to constantly check for nils, and the executor can't know
when, its definition is evaluated, whether it will be invoked via `map` or `call` (or both).
Defaulting to 0 gives the best "it just works" experience for users.