Fix the path and format for the `date` command used in the 'prototype.rb' example workflow (#490)
On MacOS with Nix, sometimes a different version of the `date` command is ending up
in $PATH when the functional tests are run. This change ensures that it is always the
expected executable when the `prototype.rb` functional test is run.

As well, different user locale settings can affect the default format printed by `date`

This PR sets both an explicit path and an explicit format when `date` is run in `prototype.rb`
to ensure that there are no spurious test failures on different machines and environments