Add optional index parameter when execution scopes are invoked with `map` (#492)
This PR adds an optional third parameter `index` to a cog_input_proc when it is
invoked inside an `execute` block that has been invoked by map. This provides functionality
equivalent to Ruby's `.map.with_index` method, and allows cogs running inside that scope to know
their index in the source enumerable.

The value of `index` will always be 0 when the scope is not invoked by `map`
(e.g., if it instead invoked by `call`). I considered having the value in this case be `nil`, but
I didn't want users to have to constantly check for nils, and the executor can't know
when, its definition is evaluated, whether it will be invoked via `map` or `call` (or both).
Defaulting to 0 gives the best "it just works" experience for users.

So, with this change, a cog's input proc can look like this at its fullest:

```
execute(:foo) do
  cog(:bar) do |my, scope_value, index|

  end
end

execute do
  map(run: :foo) { ["one", "two", "three"] }
end
```

And the `:bar` cog will be run three times with:
```
my = it's config instance
scope_value = "one"
index = 0
```

then 

```
my = it's config instance
scope_value = "two"
index = 1
```

then 

```
my = it's config instance
scope_value = "three"
index = 2
```

If a cog is run in the main executor, then `scope_value` will be `nil` and `index` will be 0. If a cog is run in a sub-executor invoked by `call` instead of `map`. then the `scope value` will be what is set on the `call`'s input, and the `index` will be 0.