Allow `call` to provide a custom index value (#493)
By default (see downstack), an executor invoked by the `call` cog will receive 0
as its option index parameter (when invoked by `map`, it will receive the index of its item
in the source enumerable).

This PR adds an optional `index` input to the `call` cog that allows the user to
specify a custom index value when invoking `call`.

I doubt this will be a common
use case, but it aligns well with the philosophy that a user should be able to
think "hmm, I wonder if it will work if I do this ... oh, cool, it does!".