Add optional parallelism to `map` cog (#496)
This PR makes it possible to use `map` to optionally process its items in parallel.

Syntax is:

```
config do
  map(:foo) do
    parallel! # allows unlimited parallelism; all items begin processing at the same time
  end

  map(:bar) do
    parallel(3) # allows up to 3 items to be processed in parallel at the same time.
  end
end
```

The items will begin executing in the order they were passed to `map`, but may complete in any order. When the output of the map is accessed, results will always be returned in the same order in which their respective inputs were given to `map`.