Reorder Config, Params, Input, Output classes in Cog definitions (#501)
We've ended up with a convention where the inner classes defined in a cog class
are given in the order Params (for system cogs), Input, Output, then Config.

This PR moves the Config class definitions to the top of each cog class.
This keeps the layout consistent across all cogs, but also puts these inner
classes in the position corresponding to which they are evaluated when the workflow
is executed.