Better nil handling for cog input objects (#500)
# Better nil handling for cog input objects

This PR introduces a (slightly) cleaner way for cogs to get validated values from their input objects (when, for instance, a part of the validation is to reject `nil`s.

__The problem:__

An input defined like this can theoretically guarantee that `name` is never `nil` when the cog is executing (`validate!` gets called by Roast when the cog's input block is executed, before the cog's `run!` method is called).
```
class Input
  #: String?
  attr_accessor :name

  def validate!
    raise Cog::Input::InvalidInputError if name.nil?
  end
end
```

But when the cog is actually running and wants to use its input values, the type checker doesn't know they can't be `nil` (or whatever) at that point. So, we either need to re-implement (some of) the validation, or do some theoretically-safe-but-gross casts. No good!

```
def execute
  actual_name = config.name  #: as String
end
```

--- 

__The improvement:__

Pull steps of the `validate!` method out into their own methods that return a valid value if their part of the input is valid, or raise an error if it isn't.

```
class Config
  #: String?
  attr_accessor :name

  def validate!
    valid_name!
  end

  def valid_name!
    raise Cog::Input::InvalidInputError if @name.nil?
    @name
  end
end
```

so now:

```
def execute
  config.valid_name!  # this is already a String
end
```

---

Just a small tweak to a pattern for now that makes a small improvement. As we build out more cogs and see more use cases, maybe a better pattern will emerge. I also think we could probably derive a clever generator for these validators, similar to the `field` helper class method on `Cog::Config`.