Add a marker for slow tests
This PR adds a `slow_test!` helper method for tests that will skip the test unless is is being run via the `:minitest_dsl_slow` rake task (or with the 'ROAST_RUN_SLOW_TESTS' environment variable set).

CI is configured to run the fast tests first, and then the slow tests, (so we'll see errors that brea
the majority of test cases quickly) and the default behaviour of `bundle exec rake` will only
run the fast tests, making it easier for quick local dev iteration.