Run cogs in custom working directory
Cogs that run external commands, like `cmd` and `agent`, or that access files on disk (like `chat`, when we give it the ability to do thing like handle image input/output) may benefit from being configured with an alternate working directory.

Note: Roast does not automatically change into the configured working directory when running
a cog for which it is specified, because doing so causes issues for parallelization.
(If two cogs are running in parallel and each attempts to change the working directory of
the main ruby process, then it gets changed for all parallel cogs, leading to unexpected behaviour).

Instead cogs that need to interact with the filesystem should use the `working_directory` config
value intentionally to operate within that folder *without* changing the working directory of the ruby
process.