Run cogs in custom working directory (#516)
Cogs that run external commands, like `cmd` and `agent`, or that access files on disk (like `chat`, when we give it the ability to do thing like handle image input/output) may benefit from being configured with an alternate working directory.

---

Note: This PR does not make Roast automatically change into the configured working directory when running a cog for which it is specified, because doing so causes issues for parallelization. (If two cogs are running in parallel and each attempts to change the working directory of
the main ruby process, then it gets changed for all parallel cogs, leading to unexpected behaviour. There's no way to effectively coördinate this.)

Instead cogs that need to interact with the filesystem, or care about their working directory for any reason, should use the `working_directory` config value internally to intentionally do whatever they want to do within that folder *without* changing the working directory of the ruby process.

(We'll maybe want to put some guardrails in place, at least a warning, to prevent the user from inadvertently changing the working directory in, say, a cog's input block. Though that's only an issue if they do use async cogs and don't realize/anticipate the side effects. Probably a niche issue.)