Add temporary mechanism for dumping all streaming claude messages (#527)
Unfortunately, there's no API documentation I can find for claude's streaming json messages, so I'm having to reverse engineer the schema and the message types from observed claude sessions. This PR adds optional functionality to dump a raw stream of claude messages to a file, so they can be replayed through the parsers and formatter to more quickly iterate on the output.