Explicitly define 'provider' config methods for `chat` cog (#551)
Tapioca (in a project that includes roast) is having difficulty with the `field` macro when generating an .rbi file for the roast gem that sorbet will be happy with.

We need to migrate to explicit method definitions for these config settings so we can add method doc comments anyway. I'm doing that in a stack of PRs, one per field, to make them easier to review.