Add convenience methods for workflow params (#562)
This PR addresses friction/annoyance in accessing the overall workflow params.

I've added methods accessible in *all* cog input contexts (in any executor scope) to access the `targets`, `args`, and `kwargs` values of the workflow params, so workflow authors do not have to worry about passing workflow params into sub-executor scopes when needed. (This is especially useful when an arg or kwargs is used to configure some specific sub-scope behaviour of the workflow but is not used by the top-level scope).

I've also added methods to access individual args and kwargs, with `?` and `!` versions for convenience, as well as a `target!` method that requires exactly one target to be present (which I think is a common use case for Roast workflows)