Add an automatic temporary directory for the workflow (#563)
This PR wraps the entire workflow execution in a `Dir.mktmpdir` block, and makes the path to the temporary directory accessible via a global convenience method similar to how the workflow params are exposed.

The workflow does not `cd` into the temporary directory. It merely creates it and cleans it up after the workflows terminates or fails.

The purpose to providing this temporary directory is to make it possible for cogs to write data to files that can be referenced by the execution of that cog (or by subsequent cogs).

There is a risk that passing state between cogs via files in the temporary directory will make the workflow harder to debug, but this functionality is something that workflow authors will implement manually if they want it, and this mechanism is much cleaner than forcing them to do that. (In particular, it's hard to manually implement a temporary directory that will be for sure cleaned up even if the workflow crashes.

As a real-world example, I'm currently writing a workflow where I need to invoke a shell command with the `cmd` cog and provide the filename of a file contained a JSON object to it as a command-line argument.

```
execute do
  cmd do
    filename = "#{tmpdir}/data.json"
    text = template("data.json.erb", { ... })
    File.write(filename, text)
    "my_command.sh -f '#{filename}'"
  end
end
```