Add .text and .lines methods to cog outputs that have textual content
These are basically just convenience methods that made it easy to get
the textual response of a cog with surrounding whitespace and newlines stripped,
and to get an array of lines with each line similarly stripped.

When a cog's output contains elements you want to process one at a time
(which happens often when a `cmd` cog produces a list of something, files, etc.),
it's common to have to split the text into lines and strip off the trailing newlines.
This makes it easier.