Provide the output, input value and loop index to the blocks given to `call`, `collect`, and `from` (#568)
If a sub-executor defines a default output with an `outputs` declaration, it can be useful to have access to that even when accessing custom output from it using `from`, `collect`, or `reduce`.

Additionally, especially when accessing the results of a running a map, it can be helpful to have the input value and input index provided for convenience.