Duplicate scope values and cog output objects to prevent users inadvertently mutating them (#577)
I'm not trying to be overly proscriptive about users passing state between cogs. It's discouraged, because it will make resuming a workflow partway through (when we implement that) probably break. But, if a user *wants* to pass global state between cogs they can always just create a global variable, or do any number of other things.

The intent here is to prevent *accidental* mutation of state that the user doesn't intend, or doesn't think about. For instance, before this change:

```
execute do
  ruby(:one) { [1,2,3] } # outputs an array of some values
  
  ruby(:two) do
    array = ruby!(:one).value
    array << 4 # mutates the array that ruby(:one) outputs, and outputs the mutated array
  end

  ruby(three) do
    value_from_one = ruby!(:one).value # this will be [1, 2, 3, 4], not [1, 2, 3]
    value_from_two = ruby!(:two).value # this will be the *same instance* as the output from :one
  end
end
```

In this toy example, the user probably won't actually be confused. But when the output objects are more complex, perhaps containing nested hashes, and the cog input logic involves extracting some piece of a cog's output or providing a default and subsequently using it somehow where it's not immediately apparent that it is a cog's output. The difference between, say, `h[:new_key] = "new_value"` and `h.merge({ new_key: "new value" })` is subtle and easy to miss.