Add file and line number information to workflow eval to improve stack trace readability (#578)
With this change, when an exception is raised inside the workflow script itself, the stacktrace emitted will actually show the filename and line number of where it was raised in the workflow script.

The rest of the stack trace is still incomprehensible, so we still have work to do there.