Access individual `map` iteration results using `from` (#588)
Building on the architectural pattern established down-stack for accessing output from a specific, individual iteration of a `repeat` loop, you can now access a specific iteration of a `map` invocation as well.

Note that the semantics here are slightly different. `map` is run on an iterable of some known size, N, and it will always produce the results of N iterations. However, because `break!` can be issued inside the map executor, some of those iterations that had not started yet will be `nil` in the map result. Attempting to access those iterations directly will raise an `IndexError`.

To help make this uncertainty easier to handle, `Map::Output` get an `iteration?` method that will check if the requested iteration has output, if you know you're working with a map that might not always complete (i.e. contains a conditional `break!`). I expect this usage will be quite uncommon for `map`, so I don't think additional shorthand is required at this time.